src: GPModel abstract class acting as an interface. Update the code and main test to this new class.

+++ src/Gpy.cpp
@@ -59,7 +59,7 @@ std::vector<std::string> Gpy::parameterNames(const std::string &object_name)
     boost::python::object gpy_ignored = boost::python::exec(gpy_str_expression, this->main_namespace);
     boost::python::object param_out = this->main_namespace["param_out"];
     std::vector<std::string> vector_out(5);
-    for (register int i = 0; i<vector_out.size(); ++i)
+    for (register size_t i = 0; i<vector_out.size(); ++i)
     {
         vector_out[i] = boost::python::extract<std::string>(param_out[i]);
     }
